initialization

+++ dtf-common/src/main/java/com/coconason/dtf/common/object/MessageType.java
@@ -3,18 +3,18 @@ package com.coconason.dtf.common.object;
 
 public interface MessageType {
 
-    Byte BUSINESS_REQ = 0;
+    int BUSINESS_REQ = 0;
 
-    Byte BUSINESS_RESP = 1;
+    int BUSINESS_RESP = 1;
 
-    Byte BUSINESS_ONE_WAY = 2;
+    int BUSINESS_ONE_WAY = 2;
 
-    Byte LOGIN_REQ = 3;
+    int LOGIN_REQ = 3;
 
-    Byte LOGIN_RESP = 4;
+    int LOGIN_RESP = 4;
 
-    Byte HEARTBEAT_REQ = 5;
+    int HEARTBEAT_REQ = 5;
 
-    Byte HEARTBEAT_RESP = 6;
+    int HEARTBEAT_RESP = 6;
 }
